+++ R/standardTables.R
@@ -20,89 +20,169 @@
 #' @param prevalence_attrition prevalence of the attrition
 #'
 #' @import here flextable dplyr
+#' @importFrom huxtable as_hux set_contents insert_row set_align everywhere
 #' @export
 table1NumPar <- function (incidence_attrition,
                           prevalence_attrition) {
 
- databaseNamePrev <- unique(prevalence_attrition$database_name)
+    if (!("reason_id" %in% names(prevalence_attrition))) {
 
-  databaseNamePrev <- databaseNamePrev[1:3]
+      prevalence_attrition <- prevalence_attrition %>%
+        mutate(reason_id = case_when(reason == "Starting population"  ~ 1,
+                                     reason == "Missing year of birth"  ~ 2,
+                                     reason == "Missing sex"  ~ 3,
+                                     reason == "Cannot satisfy age criteria during the study period based on year of birth"  ~ 4,
+                                     reason == "No observation time available during study period"  ~ 5,
+                                     reason == "Doesn't satisfy age criteria during the study period"  ~ 6,
+                                     reason == "Prior history requirement not fulfilled during study period"  ~ 7,
+                                     reason == "No observation time available after applying age and prior history criteria"  ~ 8,
+                                     reason == "Not Female"  ~ 9,
+                                     reason == "Not Male"  ~ 10,
+                                     reason == "Starting analysis population" ~ 11,
+                                     reason == "Excluded due to prior event (do not pass outcome washout during study period)" ~ 12,
+                                     reason == "Not observed during the complete database interval"  ~ 14,
+                                     reason == "Do not satisfy full contribution requirement for an interval"  ~ 16),
+               number_subjects = current_n,
+               excluded_subjects = excluded)
+
+    } else {
+
+      prevalence_attrition <- prevalence_attrition %>%
+        mutate(reason_id = case_when(reason == "Starting population"  ~ 1,
+                                     reason == "Missing year of birth"  ~ 2,
+                                     reason == "Missing sex"  ~ 3,
+                                     reason == "Cannot satisfy age criteria during the study period based on year of birth"  ~ 4,
+                                     reason == "No observation time available during study period"  ~ 5,
+                                     reason == "Doesn't satisfy age criteria during the study period"  ~ 6,
+                                     reason == "Prior history requirement not fulfilled during study period"  ~ 7,
+                                     reason == "No observation time available after applying age and prior history criteria"  ~ 8,
+                                     reason == "Not Female"  ~ 9,
+                                     reason == "Not Male"  ~ 10,
+                                     reason == "Starting analysis population" ~ 11,
+                                     reason == "Excluded due to prior event (do not pass outcome washout during study period)" ~ 12,
+                                     reason == "Not observed during the complete database interval"  ~ 14,
+                                     reason == "Do not satisfy full contribution requirement for an interval"  ~ 16))
+
+    }
+
+    # Table data
 
   tablePrevalenceAtt <- prevalence_attrition %>%
-    filter(database_name == databaseNamePrev[1]) %>%
-    group_by(step,
-             reason,
-             current_n,
-             excluded) %>%
-    summarise()
-
-  databaseNameInc <- unique(incidence_attrition$database_name)
-
-  databaseNameInc <- databaseNameInc[1:3]
-
-  tableIncidenceAtt <- incidence_attrition %>%
-    filter(database_name == databaseNameInc[1],
-           step == "Estimating incidence",
-           analysis_interval == "years") %>%
-    group_by(step,
+    group_by(reason_id,
              reason) %>%
-    summarise(current_n = round(mean(current_n), 0),
-              excluded = round(mean(excluded), 0))
+    summarise(current_n = round(mean(number_subjects ), 0),
+              excluded = round(mean(excluded_subjects ), 0)) %>%
+    mutate(analysis_step = case_when(between(reason_id, 1, 10) ~ "initial",
+                                     between(reason_id, 10, 16) ~ "prevalence")) %>%
+    filter(reason != "Do not satisfy full contribution requirement for an interval")
 
-  tablePrevIncData <- bind_rows(tablePrevalenceAtt, tableIncidenceAtt)
 
-  for (i in databaseNamePrev[2:3]) {
+  tablePrevalenceAtt <- tablePrevalenceAtt[,-1]
 
-    subPrevalenceAtt <- prevalence_attrition %>%
-      filter(database_name == i) %>%
-      group_by(step,
-               reason,
-               current_n,
-               excluded) %>%
-      summarise()
+  tablePrevalenceAtt <- tablePrevalenceAtt %>%
+    select(analysis_step, everything())
 
-    subIncidenceAtt <- incidence_attrition %>%
-      filter(database_name == i,
-             step == "Estimating incidence",
-             analysis_interval == "years") %>%
-      group_by(step,
-               reason) %>%
-      summarise(current_n = round(mean(current_n), 0),
-                excluded = round(mean(excluded), 0))
+  tablePrevalenceAtt
 
-    subPrevIncData <- bind_rows(subPrevalenceAtt, subIncidenceAtt)
+  # Incidence data
 
-    subPrevalenceAtt <- subPrevIncData[, -c(1:2)]
 
-    tablePrevIncData <- bind_cols(tablePrevIncData,
-                                  subPrevalenceAtt)
+  # incidence_attrition <- read_csv("results/resultsMock_CPRD/test database_incidence_attrition_20230427.csv")
+
+  if (!("reason_id" %in% names(incidence_attrition))) {
+
+    incidence_attrition <- incidence_attrition %>%
+      mutate(reason_id = case_when(reason == "Starting population"  ~ 1,
+                                   reason == "Missing year of birth"  ~ 2,
+                                   reason == "Missing sex"  ~ 3,
+                                   reason == "Cannot satisfy age criteria during the study period based on year of birth"  ~ 4,
+                                   reason == "No observation time available during study period"  ~ 5,
+                                   reason == "Doesn't satisfy age criteria during the study period"  ~ 6,
+                                   reason == "Prior history requirement not fulfilled during study period"  ~ 7,
+                                   reason == "No observation time available after applying age and prior history criteria"  ~ 8,
+                                   reason == "Not Female"  ~ 9,
+                                   reason == "Not Male"  ~ 10,
+                                   reason == "Starting analysis population" ~ 11,
+                                   reason == "Excluded due to prior event (do not pass outcome washout during study period)" ~ 12,
+                                   reason == "Not observed during the complete database interval"  ~ 14,
+                                   reason == "Do not satisfy full contribution requirement for an interval"  ~ 16),
+             number_subjects = current_n,
+             excluded_subjects = excluded)
+
+  } else {
+
+    incidence_attrition <- incidence_attrition %>%
+      mutate(reason_id = case_when(reason == "Starting population"  ~ 1,
+                                   reason == "Missing year of birth"  ~ 2,
+                                   reason == "Missing sex"  ~ 3,
+                                   reason == "Cannot satisfy age criteria during the study period based on year of birth"  ~ 4,
+                                   reason == "No observation time available during study period"  ~ 5,
+                                   reason == "Doesn't satisfy age criteria during the study period"  ~ 6,
+                                   reason == "Prior history requirement not fulfilled during study period"  ~ 7,
+                                   reason == "No observation time available after applying age and prior history criteria"  ~ 8,
+                                   reason == "Not Female"  ~ 9,
+                                   reason == "Not Male"  ~ 10,
+                                   reason == "Starting analysis population" ~ 11,
+                                   reason == "Excluded due to prior event (do not pass outcome washout during study period)" ~ 12,
+                                   reason == "Not observed during the complete database interval"  ~ 14,
+                                   reason == "Do not satisfy full contribution requirement for an interval"  ~ 16))
 
   }
 
+  # Table data
+
+  tableIncidenceAtt <- incidence_attrition %>%
+    group_by(reason_id,
+             reason) %>%
+    summarise(current_n = round(mean(number_subjects ), 0),
+              excluded = round(mean(excluded_subjects ), 0)) %>%
+    mutate(analysis_step = case_when(between(reason_id, 1, 10) ~ "initial",
+                                     between(reason_id, 10, 16) ~ "incidence")) %>%
+    filter(reason != "Do not satisfy full contribution requirement for an interval")
+
+
+  tableIncidenceAtt <- tableIncidenceAtt[,-1]
+
+  tableIncidenceAtt <- tableIncidenceAtt %>%
+    select(analysis_step, everything())
+
+  tableIncidenceAtt
+
+
+
+  tablePrevIncData <- union(tablePrevalenceAtt, tableIncidenceAtt)
+
+  databaseName <- unique(incidence_attrition$database_name)
+
   headerNames <- gsub("\\..*","", names(tablePrevIncData))
 
-  subtitles <- c(" ", databaseNamePrev)
+  # headerNames
 
-  flexTableAtt <- flextable(tablePrevIncData, theme_fun = theme_apa)
+  subtitles <- c(" ", databaseName)
 
-  # Deletes original header
+  subtitlesHeader <- c()
 
-  flexTableAtt <- delete_part(x = flexTableAtt, part = "header")
+  for (i in subtitles) {
 
-  # Adds header with repeaded col_keys
+    subtitlesHeader <- c(subtitlesHeader, i, " ")
 
-  flexTableAtt <- add_header_row(flexTableAtt,
-                                 values = headerNames,
-                                 colwidths = rep(1, 8),
-                                 top = FALSE)
 
-  flexTableAtt <- add_header_row(flexTableAtt,
-                                 values = subtitles,
-                                 colwidths = rep(2, length(subtitles)),
-                                 top = TRUE) %>%
-    theme_box()
+  }
 
-  return(flexTableAtt)
+  huxTableAtt <- as_hux(tablePrevIncData)
+
+  lengthNames <- length(names(huxTableAtt))
+
+  huxTableAtt <- huxTableAtt %>%
+    set_contents(1, 1:lengthNames, headerNames)
+
+  huxTableAtt <- huxTableAtt %>%
+    insert_row(subtitlesHeader, after = 0)
+
+  huxTableAtt <- huxTableAtt %>% set_align(1, everywhere, "center")
+
+  return(huxTableAtt)
+
 }
 
 #' table2IncOver
@@ -113,17 +193,7 @@ table1NumPar <- function (incidence_attrition,
 #' @export
 table2IncOver <- function (incidence_estimates) {
 
-  # incidence_estimates <- test_database_incidence_estimates_20230328 <- read_csv("results_zip/mock_results_IPCI/test database_incidence_estimates_20230328.csv")
-
-  # incidence_estimates <- incidence_estimates[grep("\\(", incidence_estimates[["outcome_cohort_name"]]), ]
-
   tableIncidence <- incidence_estimates %>%
-    # filter(n_persons %in% c(8215316,
-    #                         2283830,
-    #                         14854799,
-    #                         7310575,
-    #                         339946,
-    #                         925321)) %>%
     group_by(outcome_cohort_name,
              database_name,
              n_persons,
@@ -132,7 +202,11 @@ table2IncOver <- function (incidence_estimates) {
              incidence_100000_pys) %>%
     summarise()
 
-  return(flextable(tableIncidence))
+  tableIncidence <- as_hux(tableIncidence,
+                           add_colnames = getOption("huxtable.add_colnames", FALSE))
+
+  return(tableIncidence)
+
 
 }
 
@@ -143,23 +217,7 @@ table2IncOver <- function (incidence_estimates) {
 #' @export
 table3IncYear <- function (incidence_estimates) {
 
-  # load(here("inst/data/antibioticsProcessed/dataShiny.RData"))
-  #
-  # names(incidence_estimates)
-
-  # incidence_estimates <- incidence_estimates[grep("\\(", incidence_estimates[["outcome_cohort_name"]]), ]
-
-  # databaseNamePrev <- unique(prevalence_attrition$database_name)
-  #
-  # databaseNamePrev <- databaseNamePrev[1:3]
-
   tableIncidence <- incidence_estimates %>%
-    # filter(n_persons %in% c(8215316,
-    #                         2283830,
-    #                         14854799,
-    #                         7310575,
-    #                         339946,
-    #                         925321)) %>%
     group_by(outcome_cohort_name,
              database_name,
              incidence_start_date,
@@ -169,7 +227,10 @@ table3IncYear <- function (incidence_estimates) {
              incidence_100000_pys) %>%
     summarise()
 
-  return(flextable(tableIncidence))
+  tableIncidence <- as_hux(tableIncidence,
+                           add_colnames = getOption("huxtable.add_colnames", FALSE))
+
+  return(tableIncidence)
 }
 
 #' table4IncAge
@@ -180,23 +241,7 @@ table3IncYear <- function (incidence_estimates) {
 #' @export
 table4IncAge <- function (incidence_estimates) {
 
-  # load(here("inst/data/antibioticsProcessed/dataShiny.RData"))
-  #
-  # names(incidence_estimates)
-
-  # incidence_estimates <- incidence_estimates[grep("\\(", incidence_estimates[["outcome_cohort_name"]]), ]
-
-  # databaseNamePrev <- unique(prevalence_attrition$database_name)
-  #
-  # databaseNamePrev <- databaseNamePrev[1:3]
-
   tableIncidence <- incidence_estimates %>%
-    # filter(n_persons %in% c(8215316,
-    #                         2283830,
-    #                         14854799,
-    #                         7310575,
-    #                         339946,
-    #                         925321)) %>%
     group_by(outcome_cohort_name,
              database_name,
              denominator_age_group,
@@ -206,9 +251,10 @@ table4IncAge <- function (incidence_estimates) {
              incidence_100000_pys) %>%
     summarise()
 
-  # View(tableIncidence)
+  tableIncidence <- as_hux(tableIncidence,
+                           add_colnames = getOption("huxtable.add_colnames", FALSE))
 
-  return(flextable(tableIncidence))
+  return(tableIncidence)
 }
 
 #' table5IncSex
@@ -219,19 +265,7 @@ table4IncAge <- function (incidence_estimates) {
 #' @export
 table5IncSex <- function (incidence_estimates) {
 
-  # incidence_estimates <- incidence_estimates[grep("\\(", incidence_estimates[["outcome_cohort_name"]]), ]
-
-  # databaseNamePrev <- unique(prevalence_attrition$database_name)
-  #
-  # databaseNamePrev <- databaseNamePrev[1:3]
-
   tableIncidence <- incidence_estimates %>%
-    # filter(n_persons %in% c(8215316,
-    #                         2283830,
-    #                         14854799,
-    #                         7310575,
-    #                         339946,
-    #                         925321)) %>%
     group_by(outcome_cohort_name,
              database_name,
              denominator_sex,
@@ -241,7 +275,8 @@ table5IncSex <- function (incidence_estimates) {
              incidence_100000_pys) %>%
     summarise()
 
-  # unique(tableIncidence$n_persons)
+  tableIncidence <- as_hux(tableIncidence,
+                           add_colnames = getOption("huxtable.add_colnames", FALSE))
 
-  return(flextable(tableIncidence))
+  return(tableIncidence)
 }